+++ Packages/com.unity.inputsystem/InputSystem/Plugins/UI/InputSystemUIInputModule.cs
@@ -6,6 +6,7 @@ using UnityEngine.InputSystem.Controls;
 using UnityEngine.InputSystem.LowLevel;
 using UnityEngine.InputSystem.Utilities;
 using UnityEngine.Serialization;
+using UnityEngine.UI;
 #if UNITY_EDITOR
 using UnityEditor;
 #endif
@@ -98,6 +99,21 @@ namespace UnityEngine.InputSystem.UI
         {
             get => m_CursorLockBehavior;
             set => m_CursorLockBehavior = value;
+        }
+
+        /// <summary>
+        /// A root game object to support correct navigation in local multi-player UIs.
+        /// <remarks>
+        /// In local multi-player games where each player has their own UI, players should not be able to navigate into
+        /// another player's UI. Each player should have their own instance of an InputSystemUIInputModule, and this property
+        /// should be set to the root game object containing all UI objects for that player. If set, navigation using the
+        /// <see cref="InputSystemUIInputModule.move"/> action will be constrained to UI objects under that root.
+        /// </remarks>
+        /// </summary>
+        public GameObject localMultiPlayerRoot
+        {
+            get => m_LocalMultiPlayerRoot;
+            set => m_LocalMultiPlayerRoot = value;
         }
 
         /// <summary>
@@ -650,12 +666,15 @@ namespace UnityEngine.InputSystem.UI
                         eventData.moveVector = moveVector;
                         eventData.moveDir = moveDirection;
 
-                        ExecuteEvents.Execute(eventSystem.currentSelectedGameObject, eventData, ExecuteEvents.moveHandler);
-                        usedSelectionChange = eventData.used;
+                        if (IsMoveAllowed(eventData))
+                        {
+                            ExecuteEvents.Execute(eventSystem.currentSelectedGameObject, eventData, ExecuteEvents.moveHandler);
+                            usedSelectionChange = eventData.used;
 
-                        m_NavigationState.consecutiveMoveCount = m_NavigationState.consecutiveMoveCount + 1;
-                        m_NavigationState.lastMoveTime = time;
-                        m_NavigationState.lastMoveDirection = moveDirection;
+                            m_NavigationState.consecutiveMoveCount = m_NavigationState.consecutiveMoveCount + 1;
+                            m_NavigationState.lastMoveTime = time;
+                            m_NavigationState.lastMoveDirection = moveDirection;
+                        }
                     }
                 }
                 else
@@ -684,6 +703,45 @@ namespace UnityEngine.InputSystem.UI
                 if (!data.used && submitAction != null && submitAction.WasPressedThisFrame())
                     ExecuteEvents.Execute(eventSystem.currentSelectedGameObject, data, ExecuteEvents.submitHandler);
             }
+        }
+
+        private bool IsMoveAllowed(AxisEventData eventData)
+        {
+            if (m_LocalMultiPlayerRoot == null)
+                return true;
+
+            if (eventSystem.currentSelectedGameObject == null)
+                return true;
+
+            var selectable = eventSystem.currentSelectedGameObject.GetComponent<Selectable>();
+
+            if (selectable == null)
+                return true;
+
+            Selectable navigationTarget = null;
+            switch (eventData.moveDir)
+            {
+                case MoveDirection.Right:
+                    navigationTarget = selectable.FindSelectableOnRight();
+                    break;
+
+                case MoveDirection.Up:
+                    navigationTarget = selectable.FindSelectableOnUp();
+                    break;
+
+                case MoveDirection.Left:
+                    navigationTarget = selectable.FindSelectableOnLeft();
+                    break;
+
+                case MoveDirection.Down:
+                    navigationTarget = selectable.FindSelectableOnDown();
+                    break;
+            }
+
+            if (navigationTarget == null)
+                return true;
+
+            return navigationTarget.transform.IsChildOf(m_LocalMultiPlayerRoot.transform);
         }
 
         [FormerlySerializedAs("m_RepeatDelay")]
@@ -1909,13 +1967,15 @@ namespace UnityEngine.InputSystem.UI
         // NOTE: In the click events, we specifically react to the Canceled phase to make sure we do NOT perform
         //       button *clicks* when an action resets. However, we still need to send pointer ups.
 
-        private bool IgnoreNextClick(ref InputAction.CallbackContext context)
+        private bool IgnoreNextClick(ref InputAction.CallbackContext context, bool wasPressed)
         {
             // If explicitly ignoring focus due to setting, never ignore clicks
             if (explictlyIgnoreFocus)
                 return false;
-            // If context is cancelled (by focus change), ignore next click if device cannot run in background.
-            return context.canceled && !InputRuntime.s_Instance.isPlayerFocused && !context.control.device.canRunInBackground;
+            // If a currently active click is cancelled (by focus change), ignore next click if device cannot run in background.
+            // This prevents the cancelled click event being registered when focus is returned i.e. if
+            // the button was released while another window was focused.
+            return context.canceled && !InputRuntime.s_Instance.isPlayerFocused && !context.control.device.canRunInBackground && wasPressed;
         }
 
         private void OnLeftClickCallback(InputAction.CallbackContext context)
@@ -1925,9 +1985,10 @@ namespace UnityEngine.InputSystem.UI
                 return;
 
             ref var state = ref GetPointerStateForIndex(index);
+            bool wasPressed = state.leftButton.isPressed;
             state.leftButton.isPressed = context.ReadValueAsButton();
             state.changedThisFrame = true;
-            if (IgnoreNextClick(ref context))
+            if (IgnoreNextClick(ref context, wasPressed))
                 state.leftButton.ignoreNextClick = true;
         }
 
@@ -1938,10 +1999,11 @@ namespace UnityEngine.InputSystem.UI
                 return;
 
             ref var state = ref GetPointerStateForIndex(index);
+            bool wasPressed = state.rightButton.isPressed;
             state.rightButton.isPressed = context.ReadValueAsButton();
             state.changedThisFrame = true;
-            if (IgnoreNextClick(ref context))
-                state.leftButton.ignoreNextClick = true;
+            if (IgnoreNextClick(ref context, wasPressed))
+                state.rightButton.ignoreNextClick = true;
         }
 
         private void OnMiddleClickCallback(InputAction.CallbackContext context)
@@ -1951,10 +2013,11 @@ namespace UnityEngine.InputSystem.UI
                 return;
 
             ref var state = ref GetPointerStateForIndex(index);
+            bool wasPressed = state.middleButton.isPressed;
             state.middleButton.isPressed = context.ReadValueAsButton();
             state.changedThisFrame = true;
-            if (IgnoreNextClick(ref context))
-                state.leftButton.ignoreNextClick = true;
+            if (IgnoreNextClick(ref context, wasPressed))
+                state.middleButton.ignoreNextClick = true;
         }
 
         private bool CheckForRemovedDevice(ref InputAction.CallbackContext context)
@@ -2245,6 +2308,8 @@ namespace UnityEngine.InputSystem.UI
 
         // Navigation-type input.
         private NavigationModel m_NavigationState;
+
+        [NonSerialized] private GameObject m_LocalMultiPlayerRoot;
 
         /// <summary>
         /// Controls the origin point of raycasts when the cursor is locked.